Revert "refactoring conftest"
